bump htop to 3.0.3
newer autoconf and a few patches makes this compile.

ncurses 5.4 is older than the existence of ncurses-config, so configure
will pull it from the host system and erroneously add -ltinfo, breaking
the compilation of a test program.

ncurses is also so old that set_escdelay() doesn't exist so the old
ESCDELAY macro has to be used.

glibc is so old that -lrt must be explicitly specified.

finally, cherry-pick a commit from htop git that defines O_PATH if it's
undefined (kernels < 2.6.39).